+++ internal/survey/create.go
@@ -7,6 +7,8 @@ import (
 	"strconv"
 )
 
+const AdditionalSites = "additional_sites_name"
+
 func Create(schema *schema.CollectionInstrument, submission *s.Submission) *Survey {
 
 	lookup := substitutions.GetLookup(
@@ -30,72 +32,77 @@ func Create(schema *schema.CollectionInstrument, submission *s.Submission) *Surv
 		survey.Units = append(survey.Units, lu)
 	}
 
-	for _, lu := range GetNewUnits("additional_sites_name", submission) {
+	for _, lu := range GetNewUnits(AdditionalSites, submission) {
 		survey.Units = append(survey.Units, lu)
 	}
 
+	responseMap := submission.GetResponses()
+
 	var sections []*Section
+
 	for _, sectionTitle := range schema.ListTitles() {
 		hasAnswerValue := false
 		section := &Section{
 			Title:     substitutions.Replace(sectionTitle, lookup),
-			Instances: []*Instance{},
+			Instances: map[string]*Instance{},
 		}
-		instanceMap := map[string]*Instance{}
 		questions := schema.ListQuestionIds(sectionTitle)
 
-		for _, questionId := range questions {
-			title := substitutions.Replace(schema.GetQuestionTitle(questionId), lookup)
-			answerIds := schema.ListAnswerIds(questionId)
+		instanceCount := 0
+		instanceId := 0
+		for listItemId, data := range responseMap {
+			if listItemId == s.NonListItem {
+				instanceId = 0
+			} else {
+				instanceId = instanceCount
+				instanceCount++
+			}
 
-			for _, answerId := range answerIds {
-				answerSpecs := schema.GetAnswers(answerId)
+			instance := &Instance{
+				Id:      instanceId,
+				Answers: []*Answer{},
+			}
 
-				for _, spec := range answerSpecs {
-					answerQcode := spec.GetCode()
-					answerLabel := substitutions.Replace(spec.GetLabel(), lookup)
-					answerType := spec.GetType()
+			for _, questionId := range questions {
+				title := substitutions.Replace(schema.GetQuestionTitle(questionId), lookup)
+				answerIds := schema.ListAnswerIds(questionId)
 
-					usedInLocalUnit := false
-					for _, unit := range survey.Units {
-						//add question context to local localUnit
-						usedInLocalUnit = unit.UpdateContext(answerQcode, title, answerType, answerLabel)
-					}
+				for _, answerId := range answerIds {
+					answerSpecs := schema.GetAnswers(answerId)
 
-					responseMap := submission.GetResponses(answerQcode)
-					for _, resp := range responseList {
+					for _, spec := range answerSpecs {
+						answerQcode := spec.GetCode()
+						answerLabel := substitutions.Replace(spec.GetLabel(), lookup)
+						answerType := spec.GetType()
 
-						if resp.GetInstance() > 0 {
-							if usedInLocalUnit {
-								//this response is used in a local localUnit, so skip it
-								continue
-							}
+						usedInLocalUnit := false
+						for _, unit := range survey.Units {
+							//add question context to local localUnit
+							usedInLocalUnit = unit.UpdateContext(answerQcode, title, answerType, answerLabel)
 						}
 
-						instanceKey := strconv.Itoa(resp.GetInstance())
-						instance, found := instanceMap[instanceKey]
-						if !found {
-							instance = &Instance{
-								Id:      resp.GetInstance(),
-								Answers: []*Answer{},
-							}
-							instanceMap[instanceKey] = instance
-							section.Instances = append(section.Instances, instance)
+						if usedInLocalUnit {
+							continue
 						}
 
-						value := resp.GetValue()
+						value := data[answerQcode]
 						if value != "" {
 
 							answer := &Answer{
 								Title:    title,
 								QType:    answerType,
-								QCode:    getQCode(answerQcode, submission.GetDataVersion(), schema.GetSurveyId()),
+								QCode:    getQCode(answerQcode, schema.GetSurveyId()),
 								Label:    answerLabel,
 								Value:    value,
 								Multiple: len(answerIds) > 1 || len(answerSpecs) > 1,
 							}
 
 							instance.Answers = append(instance.Answers, answer)
+							id := strconv.Itoa(instanceId)
+							_, exists := section.Instances[id]
+							if !exists {
+								section.Instances[id] = instance
+							}
 							hasAnswerValue = true
 						}
 					}
@@ -108,97 +115,4 @@ func Create(schema *schema.CollectionInstrument, submission *s.Submission) *Surv
 	}
 	survey.Sections = sections
 	return survey
-}
-
-const LoopingDataVersion = "0.0.3"
-
-func getQCode(code, dataVersion, surveyId string) string {
-	if dataVersion == LoopingDataVersion {
-		_, err := strconv.Atoi(code)
-		if err != nil {
-			return getQCode(code[1:], LoopingDataVersion, surveyId)
-		}
-	}
-
-	if surveyId == "024" {
-		return getFuelsCode(code)
-	} else if surveyId == "194" {
-		return getRailwaysCode(code)
-	}
-
-	return code
-}
-
-func getFuelsCode(code string) string {
-	mapping := map[string]string{
-		"10":  "0a",
-		"11":  "0b",
-		"12":  "0c",
-		"13":  "0d",
-		"14":  "0e",
-		"110": "1",
-		"120": "2a",
-		"121": "2b",
-		"122": "2c",
-		"130": "3",
-		"140": "4a",
-		"141": "4b",
-		"142": "4c",
-		"150": "5",
-		"160": "6",
-		"180": "8",
-		"190": "9",
-		"200": "11",
-		"210": "12",
-		"211": "12a",
-		"220": "13",
-		"230": "15",
-		"240": "16",
-		"250": "18",
-		"260": "19",
-		"270": "20",
-		"271": "20a",
-		"280": "21",
-		"290": "23",
-		"300": "24",
-		"310": "26",
-		"320": "27",
-		"330": "28",
-		"340": "29",
-		"350": "31",
-		"360": "32",
-		"370": "34",
-		"146": "146",
-		"12a": "17",
-		"20a": "25",
-		"28":  "33",
-	}
-
-	c, found := mapping[code]
-	if found {
-		return c
-	}
-	return code
-}
-
-func getRailwaysCode(code string) string {
-	mapping := map[string]string{
-		"2":   "1.1",
-		"3":   "1.2",
-		"4":   "2.1",
-		"5":   "2.2",
-		"6":   "3.1",
-		"7":   "3.2",
-		"8":   "3.3",
-		"9":   "3.4",
-		"10":  "4.1",
-		"13":  "4.2",
-		"146": "146",
-	}
-
-	c, found := mapping[code]
-	if found {
-		return c
-	}
-	return code
 }
